poprawki po code review: 1. poprawiona nazwa pakietu (usuniecie "_)" 2. zmieniona nazwa guzika logownaia (było register) 3. hasła userow dev bez kodowania 4. poprawione wywołanie opcji BET (wysypywało się) 5. usunięcie wpisywania usera z tworzenia zakładu 6. włączenie csrf 7. poprawki checkstyle
