--- src/main/java/com/michalszekalski/bootcamp_zad26/bet/IdGenerator.java
@@ -1,10 +1,0 @@
-package com.michalszekalski.bootcamp_zad26.bet;
-
-import java.time.LocalTime;
-
-public class IdGenerator {
-
-        public static String generate(Long id) {
-            return LocalTime.now().toString().replace(":", "").substring(0, 6) + id.toString();
-        }
-    }
