+++ src/main/java/com/michalszekalski/bootcampzad26/bet/PrizeCounter.java
@@ -1,7 +1,7 @@
-package com.michalszekalski.bootcamp_zad26.bet;
+package com.michalszekalski.bootcampzad26.bet;
 
-import com.michalszekalski.bootcamp_zad26.match.Winner;
-import com.michalszekalski.bootcamp_zad26.match.Match;
+import com.michalszekalski.bootcampzad26.match.Winner;
+import com.michalszekalski.bootcampzad26.match.Match;
 
 public class PrizeCounter {
