+++ src/main/java/com/michalszekalski/bootcampzad26/user/UserRoleRepository.java
@@ -1,4 +1,4 @@
-package com.michalszekalski.bootcamp_zad26.user;
+package com.michalszekalski.bootcampzad26.user;
 
 import org.springframework.data.repository.CrudRepository;
 
@@ -6,4 +6,4 @@ import java.util.Optional;
 
 interface UserRoleRepository extends CrudRepository<UserRole, Long> {
     Optional<UserRole> findByName(String name);
- }
+}
