+++ src/main/java/com/michalszekalski/bootcampzad26/web/AdminController.java
@@ -1,9 +1,9 @@
-package com.michalszekalski.bootcamp_zad26.web;
+package com.michalszekalski.bootcampzad26.web;
 
-import com.michalszekalski.bootcamp_zad26.match.Match;
-import com.michalszekalski.bootcamp_zad26.match.MatchRepository;
-import com.michalszekalski.bootcamp_zad26.user.UserRegistrationDto;
-import com.michalszekalski.bootcamp_zad26.user.UserService;
+import com.michalszekalski.bootcampzad26.match.Match;
+import com.michalszekalski.bootcampzad26.match.MatchRepository;
+import com.michalszekalski.bootcampzad26.user.UserRegistrationDto;
+import com.michalszekalski.bootcampzad26.user.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
