+++ src/main/resources/templates/login-form.html
@@ -25,7 +25,7 @@
                 </div>
             </div>
             <div class="d-grid gap-2">
-                <button class="btn btn-primary col-12 mx-auto" type="submit">Register</button>
+                <button class="btn btn-primary col-12 mx-auto" type="submit">Login</button>
             </div>
         </div>
     </form>
